test(logstream): merge logstream writer tests
